# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for news.",
        "title": "News API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/groups": {
            "get": {
                "description": "список груп новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "список груп новин",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupsResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupsResponse500"
                        }
                    }
                }
            },
            "post": {
                "description": "Створення групи новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Створення групи новин",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "description": "news group request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewsGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            }
        },
        "/groups/{id}": {
            "get": {
                "description": "Інформація про групу новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Інформація про групу новин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid групи новин",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            },
            "put": {
                "description": "Оновлення групи новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "Оновлення групи новин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid групи новин",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "uk"
                        ],
                        "type": "string",
                        "description": "string enums",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "description": "news group request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.NewsGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            },
            "delete": {
                "description": "остаточне видалення групи новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "остаточне видалення групи новин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid групи новин",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            }
        },
        "/groups/{id}/recover": {
            "patch": {
                "description": "відновлення групи новин після м'якого видалення",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "відновлення групи новин після м'якого видалення",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid групи новин",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            }
        },
        "/groups/{id}/trash": {
            "patch": {
                "description": "м'яке видалення групи новин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "м'яке видалення групи новин",
                "parameters": [
                    {
                        "type": "string",
                        "description": "uuid групи новин",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse200"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse400"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse404"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.DocGetNewsGroupResponse500"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.FileUploadDto": {
            "type": "object",
            "properties": {
                "entity_id": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                },
                "entity_type": {
                    "type": "string",
                    "example": "news_articles"
                },
                "name": {
                    "type": "string",
                    "example": "article_img_123.png"
                },
                "path": {
                    "type": "string",
                    "example": "/uploads/article_img_123.png"
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
...
        "models.NewsGroupDTO": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string",
                    "example": "sport"
                },
                "files": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.FileUploadDto"
                    }
                },
                "published": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "example": "Спорт"
                },
                "uuid": {
                    "type": "string",
                    "example": "0194cd77-d0ab-74db-88be-f9de341a4b5f"
                }
            }
        },
...
        "request.NewsGroupRequest": {
            "type": "object",
            "required": [
                "alias",
                "title"
            ],
            "properties": {
                "alias": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "politics_ukraine"
                },
                "published": {
                    "type": "boolean",
                    "example": true
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "example": "Політика, Україна"
                }
            }
        },
...
        "response.DocGetNewsGroupResponse200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "$ref": "#/definitions/models.NewsGroupDTO"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsGroupResponse400": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsGroupResponse404": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 404
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsGroupResponse500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "щось пішло не так"
                    ]
                }
            }
        },
        "response.DocGetNewsGroupsResponse200": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.NewsGroupDTO"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "response.DocGetNewsGroupsResponse500": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 500
                },
                "data": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    },
...
}